requirements.txt
streamlit
openai
faiss-cpu
sentence-transformers
pandas
numpy